Conflicts:
- ci-core.yml: keep this branch's removal of the legacy crawler and
  series-discovery jobs (their tests were deleted here).
- ci-core.yml.backup, services/crawler/series_downloader.rs: stay deleted.
- backend/Cargo.toml: main's upgraded dev-dependency versions, plus
  this branch's wiremock.
- sec-crawler/Cargo.toml: governor/nonzero_ext stay removed (the SEC
  crawler now uses the shared HttpFetcher rate limiter).
- backend/Dockerfile: keep this branch's multi-target Dockerfile, with
  RUST_VERSION 1.98.1 on slim-bookworm; bump the crawler clippy job's
  toolchain to 1.98.1 too.

Adapt the new crawler code to the upgraded crates:
- diesel-async 0.9: transactions take async closures (scoped_futures is
  gone); the point-count helper becomes an async fn because the
  get_result future now borrows the connection.
- rand 0.10: rng()/random_range/random_bool/random via RngExt.

+++ backend/crates/econ-graph-core/src/models/data_point/tests.rs
@@ -44,7 +44,7 @@ mod simple_tests {
 
         let change = yoy_change.unwrap();
         // Should be 10% change
-        assert!(change > BigDecimal::from(9) && change < BigDecimal::from(11));
+        assert!(change > 9 && change < 11);
     }
 
     #[test]
@@ -72,7 +72,7 @@ mod simple_tests {
 
         let change = qoq_change.unwrap();
         // Should be 5% change
-        assert!(change > BigDecimal::from(4) && change < BigDecimal::from(6));
+        assert!(change > 4 && change < 6);
     }
 
     #[test]
@@ -100,7 +100,7 @@ mod simple_tests {
 
         let change = mom_change.unwrap();
         // Should be 2% change
-        assert!(change > BigDecimal::from(1) && change < BigDecimal::from(3));
+        assert!(change > 1 && change < 3);
     }
 
     #[test]
